scripts/9/htusocialnetwork/requirements.txt
django==3.2
djangorestframework==3.13.1